LPS-93653 Migrate portal-remote-json-web-service-extender-test to use arquillian bridge.
