+++ modules/apps/portal-remote/portal-remote-json-web-service-extender-test/bnd.bnd
@@ -1,5 +1,3 @@
-Bundle-Activator: com.liferay.portal.remote.json.web.service.extender.test.internal.WebServiceExtenderTestBundleActivator
 Bundle-Name: Liferay Portal Remote JSON Web Service Extender Test
 Bundle-SymbolicName: com.liferay.portal.remote.json.web.service.extender.test
 Bundle-Version: 4.0.0
-Private-Package: com.liferay.portal.remote.json.web.service.extender.test.internal
