+++ modules/apps/portal-remote/portal-remote-json-web-service-extender-test/build.gradle
@@ -1,9 +1,5 @@
 dependencies {
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
-
-	testIntegrationCompile group: "com.liferay.arquillian", name: "com.liferay.arquillian.arquillian-container-liferay", version: "1.0.4"
-	testIntegrationCompile group: "com.liferay.arquillian", name: "com.liferay.arquillian.arquillian-deployment-generator-bnd", version: "1.0.4"
-	testIntegrationCompile group: "org.apache.httpcomponents", name: "httpclient", version: "4.3.5"
-	testIntegrationCompile group: "org.jboss.arquillian.junit", name: "arquillian-junit-container", version: "1.1.10.Final"
+	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
+	testIntegrationCompile group: "org.osgi", name: "org.osgi.core", version: "6.0.0"
+	testIntegrationCompile project(":test:arquillian-extension-junit-bridge")
 }
