--- modules/apps/portal-remote/portal-remote-json-web-service-extender-test/src/main/java/com/liferay/portal/remote/json/web/service/extender/test/internal/TestWebService.java
@@ -1,29 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.remote.json.web.service.extender.test.internal;
-
-import com.liferay.portal.kernel.jsonwebservice.JSONWebService;
-
-/**
- * @author Carlos Sierra Andrés
- */
-@JSONWebService
-public class TestWebService {
-
-	public int sum(int a, int b) {
-		return a + b;
-	}
-
-}
