--- modules/apps/portal-remote/portal-remote-json-web-service-extender-test/src/main/java/com/liferay/portal/remote/json/web/service/extender/test/internal/WebServiceExtenderTestBundleActivator.java
@@ -1,49 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.remote.json.web.service.extender.test.internal;
-
-import java.util.Dictionary;
-import java.util.Hashtable;
-
-import org.osgi.framework.BundleActivator;
-import org.osgi.framework.BundleContext;
-import org.osgi.framework.ServiceRegistration;
-
-/**
- * @author Raymond Augé
- */
-public class WebServiceExtenderTestBundleActivator implements BundleActivator {
-
-	@Override
-	public void start(BundleContext bundleContext) throws Exception {
-		Dictionary<String, Object> properties = new Hashtable<>();
-
-		properties.put("json.web.service.context.name", "test");
-		properties.put("json.web.service.context.path", "webservice");
-
-		_serviceRegistration = bundleContext.registerService(
-			Object.class, new TestWebService(), properties);
-	}
-
-	@Override
-	public void stop(BundleContext bundleContext) throws Exception {
-		_serviceRegistration.unregister();
-
-		_serviceRegistration = null;
-	}
-
-	private ServiceRegistration<Object> _serviceRegistration;
-
-}
